requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
numpy
# Dependências do LLM/Vector DB
google-genai
qdrant-client
# Dependência do DB Principal (SQLite via SQLAlchemy)
SQLAlchemy
tqdm